host side cleanup
