+++ hosted/src/lib.rs
@@ -1,1 +1,5 @@
+#![feature(ascii_char)]
+
 pub mod spi;
+#[cfg(feature = "tui")]
+pub mod tui;
